07_jwt_all_included/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
bcrypt
